requirements.txt
torch
numpy
pillow
opencv-python
matplotlib
